update health bar and damage calculation

+++ Assets/ProjectFiles/Scripts/CharacterInfo/ICharacterInfo.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public interface ICharacterInfo
 {
-    public int GetHealth();
-    public void SetHealth(int amount);
-    public int GetAttack();
+    public float GetHealth();
+    public void SetHealth(float amount);
+    public float GetAttack();
     public string GetName();
     public Transform GetTransform();
     public string GetCharType();
